requirements.txt
openstacksdk==0.26.0
pyVim==2.0.24
pyVmomi==6.7.1.2018.12
